Log Recorder interface
